Rename TooltipData to TooltipTable
